+++ lib/infrastructure/ui/dialog_impl.dart
@@ -17,19 +17,22 @@ import 'package:get_arch_quick_start/quick_start_part.dart';
 ///
 /// 本文件依赖于 bot_toast 包
 
+/// 接收Failure, 构建Dialog
+typedef FailureDialogBuilder = Widget Function(BuildContext ctx, Failure f);
+
 ///
 /// 内部存储一个Failure与Dialog的Map
 ///   例如某处抛出 "未登录Failure", 那么调用[map]中对应的Failure的Dialog
 ///
-/// 这里推荐使用 [QuickAlter]或其继承类作为 Value,
+/// 这里推荐使用 [QuickAlert]或其继承类作为 Value,
 ///
-/// [map] <错误类型, 错误所对应的Dialog>
+/// [map] <Failure类型, Failure所对应的Dialog>
 ///   Dialog内应当有对应的Failure所应当展示的信息内容,
 ///   如果有跳转页面需求, 还需要为Dialog按钮的点击回调中
 ///     添加对应页面跳转逻辑(例如"未登录错误Dialog"可以跳转到 "登录页")
 //@lazySingleton 请使用 module或者手动注册本类
 class FailureRoute {
-  final Map<Failure, Widget> map;
+  final Map<Type, FailureDialogBuilder> map;
 
   FailureRoute(this.map);
 }
@@ -42,8 +45,8 @@ class FailureRoute {
 /// [customActions] 与 [onCancel],[onConfirm]互斥
 /// [onConfirm] "确认"按钮回调
 /// [onCancel]  "取消"按钮回调
-class QuickAlter extends AlertDialog {
-  QuickAlter({
+class QuickAlert extends AlertDialog {
+  QuickAlert({
     Widget title: const Text('提示'),
     Widget content,
     List<Widget> customActions,
@@ -81,8 +84,9 @@ class QuickAlter extends AlertDialog {
         );
 }
 
-//@prod // 手动注册
-//@LazySingleton(as: IDialog)
+@prod
+@test
+@LazySingleton(as: IDialog)
 class QuickDialog extends IDialog {
   static QuickDialog _instance;
 
@@ -90,6 +94,8 @@ class QuickDialog extends IDialog {
     if (_instance == null) _instance = QuickDialog();
     return _instance;
   }
+
+  static QuickDialog get I => instance;
 
   ///
   /// [T] 表示pop后返回值的类型
@@ -101,7 +107,7 @@ class QuickDialog extends IDialog {
   /// [customActions] 自定义Action,
   ///   值为null时, Dialog将会有两个默认的"确认"和"取消"按钮,
   ///   值为[] 时, 将隐藏 action按钮
-  Future<T> selectTipsWithCtx<T>({
+  Future<T> selectTips<T>({
     @required BuildContext ctx,
     String title: '提示',
     @required Widget content,
@@ -116,7 +122,7 @@ class QuickDialog extends IDialog {
 
     return showDialog<T>(
         context: ctx,
-        builder: (ctx) => QuickAlter(
+        builder: (ctx) => QuickAlert(
               title: Text('$title'),
               content: content,
               onConfirm: () {
@@ -136,15 +142,15 @@ class QuickDialog extends IDialog {
   }
 
   ///
-  /// 尽量使用 [selectTipsWithCtx] 其性能更好
-  selectTips({
+  /// 尽量使用 [selectTips] 其性能更好
+  selectTipsWithoutCtx({
     String title: '提示',
     @required Widget content,
     @required VoidCallback onConfirm,
     VoidCallback onCancel,
   }) =>
       BotToast.showWidget(
-          toastBuilder: (cancelFunc) => QuickAlter(
+          toastBuilder: (cancelFunc) => QuickAlert(
                 title: Text('$title'),
                 content: content,
                 onConfirm: () {
@@ -159,24 +165,34 @@ class QuickDialog extends IDialog {
 
   ///
   /// 异常提示
+  /// 请在View层使用Dialog, 不要在ViewModel或其他地方使用Dialog!
+  ///
   /// 在DI中注册[FailureRoute]后,将会弹出Failure对应的Dialog
-  err(dynamic failure, {BuildContext ctx, dynamic tag}) {
+  /// [ctx] 不能为空,否则弹出的Dialog无法关闭.
+  ///   (当然你也可以实现一个能够在无ctx状态下弹出Dialog并能正常关闭的方案, 但非常不建议这样做)
+  /// [tag] 错误标签, 便于Debug
+  err(dynamic failure, {@required BuildContext ctx, dynamic tag}) {
+    assert(ctx != null, 'err中的ctx不能为null!,否则Dialog将无法关闭');
     if (failure is! Failure)
       failure = FeedBackUnknownFailure(failure.toString(), tag ?? '来自Dialog');
-    final r = GetIt.I.isRegistered<FailureRoute>();
-    if (r) {
-      if (!kReleaseMode) print('QuickDialog.err # 您尚未注册"FailureRoute"!');
-    } else {
-      final dialog = GetIt.I<FailureRoute>().map[failure];
-      if (dialog != null) {
-        return widget(ctx: ctx, dialog: dialog);
+
+    if (GetIt.I.isRegistered<FailureRoute>()) {
+      final buildDialogFunc = GetIt.I<FailureRoute>().map[failure.runtimeType];
+      if (buildDialogFunc != null) {
+        if (!kReleaseMode && ctx == null)
+          print(
+              '\n\nQuickDialog.err# tag[$tag]未添加BuildContext,可能无法正常路由跳转! Failure[$failure]\n\n');
+        return widget(
+            ctx: ctx, dialogBuilder: (c) => buildDialogFunc(c, failure));
       }
+    } else {
+      if (!kReleaseMode) print('QuickDialog.err # 您尚未注册"FailureRoute"!');
     }
     _onNeedFeedback(failure);
   }
 
   static _onNeedFeedback(Failure f) => BotToast.showWidget(
-      toastBuilder: (cancelFunc) => QuickAlter(
+      toastBuilder: (cancelFunc) => QuickAlert(
             title: Text(f.runtimeType.toString()),
             content: Text(f.msg),
             customActions: <Widget>[
@@ -192,53 +208,41 @@ class QuickDialog extends IDialog {
   @override
   toast(String s, {BuildContext ctx}) => ctx != null
       ? snack(
-          ctx: ctx, content: Text('$s'), behavior: SnackBarBehavior.floating)
+          SnackBar(content: Text('$s'), behavior: SnackBarBehavior.floating),
+          ctx: ctx)
       : BotToast.showText(text: s);
 
   @override
-  snack({
-    @required BuildContext ctx,
-    @required Widget content,
-    Color backgroundColor,
-    double elevation,
-    ShapeBorder shape,
-    SnackBarBehavior behavior,
-    SnackBarAction action,
-    Duration duration,
-    Animation<double> animation,
-    VoidCallback onVisible,
-  }) =>
-      Scaffold.of(ctx).showSnackBar(SnackBar(
-        content: content,
-        backgroundColor: backgroundColor,
-        elevation: elevation,
-        shape: shape,
-        behavior: behavior,
-        action: action,
-        duration: duration,
-        animation: animation,
-        onVisible: onVisible,
-      ));
+  snack(SnackBar snackBar, {BuildContext ctx}) =>
+      Scaffold.of(ctx).showSnackBar(snackBar);
 
-  /// [ctx] 可选, 一般推荐加上,[ctx]为空则会使用BotToast
+  /// [ctx] 可选, 添加后性能更好,[ctx]为空则会使用BotToast
   ///
   /// showDialog返回值是 [T],即[dialog]内部通过路由pop()返回的值
   ///
   /// BotToast返回的是 CancelFunc, 调用后可以关闭Dialog,
   ///   如果需要对[dialog]内的操作做出反应,请使用回调函数
   @override
-  widget<T>({BuildContext ctx, Widget dialog}) => ctx == null
-      ? BotToast.showWidget(toastBuilder: (_) => dialog)
-      : showDialog<T>(context: ctx, builder: (ctx) => dialog);
+  widget<T>({
+    @required BuildContext ctx,
+    WidgetBuilder dialogBuilder,
+    bool barrierDismissible = true,
+  }) =>
+      ctx == null
+          ? BotToast.showWidget(toastBuilder: (_) => dialogBuilder(null))
+          : showDialog<T>(
+              context: ctx,
+              builder: dialogBuilder,
+              barrierDismissible: barrierDismissible);
 }
 
-//@test // 手动注册
-//@LazySingleton(as: IDialog)
-class TestDialog extends IDialog {
+@dev
+@LazySingleton(as: IDialog)
+class DevQuickDialog extends IDialog {
   @override
   err(dynamic failure, {dynamic tag, BuildContext ctx}) {
     print('''
-╔═╣ TestDialog.err | Tag: $tag╠══╗
+╔═╣ DevDialog.err | Tag: $tag╠══╗
   $failure
 ╚══════════════════════════════════════╝
     ''');
@@ -247,40 +251,50 @@ class TestDialog extends IDialog {
   @override
   toast(String s, {BuildContext ctx}) {
     print('''
-╔═╣ TestDialog.text ╠═╗
+╔═╣ DevDialog.text ╠═╗
   $s
 ╚═══════════════════════════╝
     ''');
   }
 
   @override
-  snack(
-      {BuildContext ctx,
-      Widget content,
-      Color backgroundColor,
-      double elevation,
-      ShapeBorder shape,
-      SnackBarBehavior behavior,
-      SnackBarAction action,
-      Duration duration,
-      Animation<double> animation,
-      onVisible}) {
+  snack(SnackBar snackBar, {BuildContext ctx}) {
     print('''
-╔═╣ Test/DevDialog.snack ╠═╗
-  ${content.toStringDeep()}
+╔═╣ DevDialog.snack ╠═╗
+  ${snackBar.toStringDeep()}
 ╚═══════════════════════════╝
     ''');
   }
 
   @override
-  widget<T>({BuildContext ctx, Widget dialog}) {
+  widget<T>(
+      {BuildContext ctx,
+      WidgetBuilder dialogBuilder,
+      bool barrierDismissible = true}) {
     print('''
-╔═╣ Test/DevDialog.widget ╠═╗
+╔═╣ DevDialog.widget ╠═╗
   Context: [$ctx]
-  ${dialog.toStringDeep()}
+  barrierDismissible: [$barrierDismissible]
+  ${dialogBuilder.call(ctx).toStringDeep()}
 ╚═══════════════════════════╝
     ''');
   }
-}
 
-class DevDialog extends TestDialog {}
+  @override
+  Future<T> selectTips<T>(
+      {BuildContext ctx,
+      String title = '提示',
+      Widget content,
+      onConfirm,
+      onCancel,
+      List<Widget> customActions}) {
+    print('''
+╔═╣ DevDialog.selectTipsWithCtx ╠═╗
+  Context: [$ctx]
+  title: [$title]
+  content: [$content]
+╚═══════════════════════════╝
+    ''');
+    return null;
+  }
+}